# Faster MemoryPoolIterator2.CopyFrom