+9.5% RPS

Before
> Running 30s test @ http://.../plaintext
  32 threads and 1024 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     7.36ms   17.09ms 341.62ms   96.60%
    Req/Sec    62.36k    11.00k   95.84k    92.01%
  59601197 requests in 30.03s, 7.33GB read
  Socket errors: connect 35, read 0, write 0, timeout 0
Requests/sec: 1984923.91
Transfer/sec:    249.87MB

After 
> Running 30s test @ http://.../plaintext
  32 threads and 1024 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     7.47ms   18.58ms 313.68ms   96.10%
    Req/Sec    70.66k     7.23k  328.02k    90.11%
  65455288 requests in 30.10s, 8.05GB read
  Socket errors: connect 35, read 0, write 0, timeout 1
Requests/sec: 2174662.49
Transfer/sec:    273.76MB